Add cabal flag for static compilation
Adds a flag to the hadolint package that toggles static compilation.
By default I set it to `false` and enable this flag in the
`Dockerfile`, `appveyor.yml` and `travis.yml`.

The reason is that some systems don't easily support static linking,
so using the flag when you want it is nicer instead of having to patch
around this problem.