healthcheck: simplify rule, fix edge case
- Inline two helper functions
- Remove a case from the (From BaseImage {image, alias = Nothing})
  pattern

The removed case was consolidated with the following case, because it
would be incorrect to not trigger when the build stage has no alias set
and has no `HEALTHCHECK` inside the stage, regardless of whether or not
a stage is following. This is also expressed in the new test cases.
In case of a Dockerfile used to build multiple resulting images, all
stages without an alias must have or inherit a `HEALTHCHECK`, not just
the last one.