dl3044: fix bug with variable detection
When a variable is defined in terms of another variable and that other
variable is not referred to with curly-baces notation, the definition of
the first variable is split on the variable name. This results in a list
of remaining strings. The beginnings of the remaining strings is used to
determine, if the variable name that was found was just a substring of
another variable name, or is the actual variable.
Since the `Text.splitOn` also returns the first part before the
variable, it is wrongfully determined, that the variable was found in
cases where its name was just a substring in another variable name.
Dropping this substring that never could begin with a terminator for a
variable name therefore fixes this problem.

fixes #568