label-schema: proper parsing for label type
- Put label type into its own sum data type
- Add parsers capable of parsing `Text.Text` into `LabelType`,
  both for yaml files as well as commandline arguments.

This improves behaviour when specifying a label schema. Before this
change, invalid label types would be ignored silently and treated as if
they were type `text`. Now, if given, the label type must be one of
`url`, `spdx`, `hash`, `rfc3339`, `semver` or `text`. Not giving a label
type is allowed and will also produce the same behaviour as `text`. But
any other type will produce a more or less helpful error message and
abort the program.