tests: Split tests into dedicated files (#582)
Given how humongously large Spec.hs has grown, it also had become harder
to handle and generally just a bear to work with. It also became
increasingly hard to determine whether or not a particular rule was
tested against a particular example, as sometimes these tests were
strewn about in different sections of the file.

It is still ok for some tests to be directly in Spec.hs as they are
concerned with the general functionality of the program, like testing
inline ignores. Other tests - especially for individual rules - are
moved to their individual files, giving a much easier access for
developers.