options: introduce failure severity threshold
- add new option `--failure-threshold`/`-t` for specifying a failure
  threshold
- add new config option `failure-threshold` to the config file
- add schemastore schema under `contrib/` directory

The specified failure threshold (info by default) will cause hadolint to
not return with a failure code when only rules with a lesser severity
 are violated.
Accepted values are `error`, `warning`, `info`, `style`, `ignore` and
`none`, where `info` is the default and `ignore` and `none` both cause
Hadolint to exit with an error code if any rule is violated. A threshold
of level `error` has the same effect as passing `--no-fail`.

As discussed in https://github.com/hadolint/hadolint/issues/602, the
schemastore schema should be self hosted. This change to the config file
(and thus to the schemastore schema as well) is the perfect opportunity
to do so

fixes: #607
fixes: #602