config: fix parsing problems
Parsing the contents of the `failure-threshold` property was done
improperly and resulted in all results being discarded in case the
property was absent. The observed behaviour was seemingly the same as if
the config file had not been found, but in fact it was a problem of the
parser.

- Make all severity overrides more solid as types by implementing the
  to-be-overridden rules as rule codes instead of `Text.Text`s
- Make `DLSeverity` an instance of `Yaml.FromYAML`, `Semigroup` and
  `Monoid` to properly handle parsing from yaml files and combining
  multiple severities.

With `DLSeverity` being a more sophisticated type and implementing its
own yaml parser, parsing the config and combining it with the
configuration obtained from commandline options has much more consistent
behaviours with respect to the existing configuration properties.

fixes: #624