DL3010: Allow intentionally unextracted archives
- Add advanced logic to DL3010 to detect archive extraction.
- Add more specific tests for DL3010

Sometimes it is desired to copy an archive into a Docker image. In these
cases, DL3010 should not trigger. This can be detected by searching for
commands extracting an archive that was added by a `COPY` instruction
earlier.
The new behaviour triggers only if an archive is `COPY`ed into the image
_and_ extracted (e.g. by `RUN tar -xf archive-foo.tar.gz`).

fixes: #709