DL3059: ignore chains of commands
- Make DL3059 ignore consecutive `RUN` instructions if they contain
  chained-together commands.
- Add tests

Multiple consecutive `RUN` instruction usually are a sign of trouble in
a Dockerfile. But when the developer has chained together multiple
commands in a `RUN` instruction and still uses multiple consecutive
`RUN` instructions, it can be assumed that this is a consious choice and
thus DL3059 should ignore this.
This behaviour helps to quell DL3059s tendency to force developers to do
'cache busting' when not necessary but will still catch cases where each
command is mindlessly put into its own `RUN` instruction.

fixes: #715