Overhaul: Split Configuration data types
The configuraion data is used in two stages. An 'assembly' stage at the
beginning and a 'read-only' stage during the actual linting and output.

Split the `Configuration` data type into a `PartialConfiguration` and a
final `Configuration` data type. The partial type may have unset fields
or missing information but the actual `Configuration` data type always
has all fields.

This allows for some simplification in the later stages of the program
because when given a `Configuration` now it always has all fields and
it is therefore guaranteed safe to read all fields.
Similarly the possibility of incomplete data can easily be handled with
the `PartialConfiguration` data type and the convnience function
`applyPartialConfiguration` can be used to construct an actual
`Configuration` given a default `Configuration` and a
`PartialConfiguration`.

One result is that the linting step may never run unconfigured, but it
is not hard to just give it a default configuration.