Overhaul: Reorganize tests, redesign configuration (#721)
* Overhaul: Reorganize tests, redesign configuration

config: redesign configuration handling

- Redesign configuration handling to have a clear order of precedence
  default < environment < config file < comandline options
- Implement more configuration options as environment variables and
  as options in the config files
- Define a clear default configuration
- Move configuration code into its own submodule and keep app/Main.hs
  clear of it
- Make configuration parsing and handling testable
- Add sensible locations to search for config files.
- Reorganize code and optimize for cleaner imports

tests: Use Hspec discovery to auto-discover tests

- Replace boilerplate code with hspec discovery.
  This also enables the use of Hspec hooks.
- Rename test submodules to conform with Hspecs expectations
- Make all fields in the config optional thus enabling proper priority
  based override behaviour.
- Add tests for config parsing and configuration handling
- Add tests for output formats
- Add tests for commandline argument parsing

other:

- Version bump language-docker 10.1.2 --> 10.2.0
- Verbose output now shows the whole running configuration
- Convert custom read functions for label types and sevrities to generic
  instance of `Read`.

fixes: #716
fixes: #570

* Overhaul: Replace unsafe and unnecessary code

- replace `fromJust` with pattern matching or proper functions like
  `fromMaybe`
- give descriptive names to functions with confusing or placeholder
  names
- replace custom instances of `Read` with reader functions
- replace custom instances of `Show` with instances of `Pretty`
- replace `<>>` operator with `<|>` from `Control.Applicative`

* Overhaul: Flatten config data structure

- Merge the different nested configuration data structures
- Simplify constructor usage in test suite

Instead of three different configuration/options data structures nested
within each other, now only one remains (with the exception of the
invokation options on the command line. This massively simplifies
constructor usage and also makes it much easier to access different
configuration options from all parts of the program.
Furthermore, the configuration data structures are no longer distributed
accross three different independant modules and pollute modules not
tasked with configuration with their code.

* Overhaul: beautify the config pretty printer

Beautify the pretty printer of the configuration. Utilizing more of
built in functions of the pretty printer, the output becomes less
fragile and better understandable.

- remove "\n" hacks for newlines with the pretty printer
- use the correct concatenation operator
- use nesting functions instead of manually adding spaces

* Overhaul: Improve formatter tests

- Consolidate some boilerplate code into a helper function
- Add Sarif formatter tests

* Overhaul: Split Configuration data types

The configuraion data is used in two stages. An 'assembly' stage at the
beginning and a 'read-only' stage during the actual linting and output.

Split the `Configuration` data type into a `PartialConfiguration` and a
final `Configuration` data type. The partial type may have unset fields
or missing information but the actual `Configuration` data type always
has all fields.

This allows for some simplification in the later stages of the program
because when given a `Configuration` now it always has all fields and
it is therefore guaranteed safe to read all fields.
Similarly the possibility of incomplete data can easily be handled with
the `PartialConfiguration` data type and the convnience function
`applyPartialConfiguration` can be used to construct an actual
`Configuration` given a default `Configuration` and a
`PartialConfiguration`.

One result is that the linting step may never run unconfigured, but it
is not hard to just give it a default configuration.