DL3049: Fix behaviour with multistage builds
- Fix DL3049 behaviour with multistage builds

In a multistage build, when a stage has an alias, it is reasonable to
assume that it will not be the last stage in the build process. Rather
there will be another stage inheriting or copying from that stage.
Therefore required labels don't necessarily need to be present in the
stage with the alias.

fixes: #758
Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>