DL3061: Check for instruction order (#763)
Not all dockerfiles are valid just because all instructions in the
dockerfile are individually valid. For example, a `RUN` instruction may
not preceed the first `FROM` instruction and so on. The new rule DL3061
checks that the order of the instructions is not completely bogus. That
means the Dockefile must begin with a comment, an `ARG` or a `FROM`
instruction.

- Add new rule DL3061 to check for instruction order
- Add tests for rule DL3061
- Fix regression tests wrt. rule DL3061

fixes: #737