config: respect NO_COLOR specification
- Decide whether to print ANSI color codes or not based on the
  _presence_, not the value of the NO_COLOR environment variable

The specification provided by https://no-color.org regarding the
interpretation of the NO_COLOR environment variable specifically state
that the value does not matter, only whether it is set or not should
change program behavior.
Therefore, Hadolint should respect that and ignore the value.

There is a platform specific distinction in the test suite because on
Windows it is expected that setting an environment variable to the empty
string is equal to unsetting it, in contrast to POSIX platforms where an
environment variable can be set to the empty string and still count as
set. With $NO_COLOR it only matters if it is set, not what it is set to,
the case on POSIX where it is set to the empty string should be
respected.