DL3057: fix multiple warnings for multistage builds (#851)
DL3057 checks that images produced in multistage builds have a
healthcheck (if the rule is enabled). But if one stage already has the
healthcheck instruction, the other stages should not trigger a warning.
If the stage comes earlier, it's obviously not used as a build result,
therefore doesn't need a healthcheck instruction on its own.
If the stage comes later, it will inherit the healthcheck instruction
from the stage that has it.

This fix works because when a healthcheck is found in a build stage,
that stage is marked as good, but also all stages it inherited from are
also marked as good. Only stages that don't inherit the healthcheck
instruction remain marked as bad, if they don't have the instruction
themselves. Therefore Dockerfiles, which produce multiple images are
also considered.

fixes: #843
Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>