Pragma: Allow comments following inline (#1059)
Allow comments and whitespace to follow inline with pragma comments,
e.g.:

```
# hadolint ignore=DL3006  # comment on why to ignore this rule here
# hadolint shell=powershell  
```

Modify the parsers for pragma comments to ignore both:

1) just whitespaces following the inline comment.
2) whitespaces followed by `#`, and anything after until end of line.

This improves robustness of the parser and allows users to add helpful
comments to their pragmas.
Note that not separating comment from pragma with an additional `#` will
result in a silent failure to parse and the pragma will have no effect.

fixes: hadolint/hadolint#997
fixes: hadolint/hadolint#1028

Signed-off-by: Moritz Röhrich <moritz@ildefons.de>