use docker buildx to create manifests
By default Docker creates provenance images, i.e. the object in the
registry is a manifest index with both image data and provenance data.
To create a multi-arch image from that, it's necessary to use buildx
imagetools instead of the docker manifest command, because it can handle
these manifest indices.

Signed-off-by: Moritz Röhrich <moritz@ildefons.de>