DL3041, DL3033: Handle RPM package epoch
In Fedora, RHEL and similar distributions, packages may have an epoch
prefixed to their version. This is an integer separated by a colon from
the actual version number. The epoch is used to make all packages
sortable, regardless of whether the upstream project uses sortable
version identifiers.
To deal with this, in Hadolint we must accept that a dnf or yum package
may contain a colon in their version number.

related-to: hadolint/hadolint#1117

Signed-off-by: Moritz Röhrich <moritz@ildefons.de>