DL3009: Allow either cache or tmpfs mounts
Allow apt cache and lists directories to be either cache or tmpfs mounts
without warning.

DL3009 warns if apt cache or list data is kept unnecessarily in the
container image, as this increases the image size.
There are several ways to avoid this: 1) remove the cache and lists
directories after use, 2) mount a cache from the host to these
directories and 3) mount a tmpfs from the host to the directories.
This change modifies the DL3009 rule such that it won't trigger a
warning if the a mount of type tmpfs is used for either the cache or
lists directory.

related-to: hadolint/hadolint#1095

Signed-off-by: Moritz Röhrich <moritz@ildefons.de>