dl3062: deal with tool commands
DL3062: Deal with `go ... tool` commands. Since Golang >= 1.24 the
package tooling for Golang can also manage tools as dependencies.
This introduces variations for the `go install` and `go get` commands.
The additional `tool` and `-tool` arguments respectively have to be
accepted by DL3062 and shall not lead to false positives.

related-to: hadolint/hadolint#1132

Signed-off-by: Moritz Röhrich <moritz@ildefons.de>